genai-ecom-agent/genai-ecom-agent/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pandas
plotly
